@@ -1,8 +1,8 @@
 package com.dertyp7214.changeloglibexampleapp;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.widget.Toast;
 
 import com.dertyp7214.changelogs.ChangeLog;
